build: Fix installation of neovim in setup script

+++ setup.sh
@@ -52,20 +52,22 @@ if !isInstalled build-essential; then
 fi
 
 #install neovim
-if [[ ! -d $builddir/nvim ]]; then
-	if [[ $(git clone https://github.com/jibon0070/nvim-kick-starter.git nvim) ]]; then
-		echo "Failed to clone nvim"
+if ! isInstalled nvim; then
+	if [[ ! -d $builddir/nvim ]]; then
+		if [[ $(git clone https://github.com/jibon0070/nvim-kick-starter.git nvim) ]]; then
+			echo "Failed to clone nvim"
+			exit 1
+		fi
+	fi
+	cd $builddir/nvim
+	./setup.sh
+	if [[ $? -ne 0 ]]; then
+		echo "Failed to install nvim"
 		exit 1
 	fi
+	cd $builddir
+	rm -rf nvim
 fi
-cd $builddir/nvim
-./setup.sh
-if [[ $? -ne 0 ]]; then
-	echo "Failed to install nvim"
-	exit 1
-fi
-cd $builddir
-rm -rf nvim
 
 #install nodejs
 nala install nodejs npm -y
